Fix regression with drag points
Added the check for the mouse changing position before initiating drag
again. Was removed for testing for the new method of getting mouse
position.

Also added a check to the head drag point for select mode to prevent the
maid animation from stopping. Selecting the maid's head doesn't change
the pose so stopping the animation is unnecessary.