Rework EnvironmentManager
Fixes an issue where flipping the background upside down distorts
skirts.

Needed to attach the drag point to the actual bg object but it changes
every time the bg changes so a patcher was made to listen for bg
changes. This allows for bg changes to be detected even from other
sources like the bg panel in studio mode.