Fix camera z rotation slider resetting other axes
Also update Z rotation slider when resetting the camera with Q + R

This was a left over bug from when I was experimenting with rotation
sliders for all axes.

Don't know how I didn't notice that the camera rotation was resetting
for axes other than Z while testing but whatever.