Fix drag point misalignment from maid body
Arm and finger bones are misaligned from the actual maid's body when the
maid's body is heavily modified through modslider. Modslider can be used
to make maids body parts much larger and smaller which affects the
position of the arm and finger bones.

This commit changes the drag point's target position from the bones to
the maid's body instead which results in the drag points aligning with
the maid's body perfectly.

These changes are very experimental and dirty.