Fix non-existent custom pose breaking scene load
Original method of checking if the pose is custom was to check if the
pose name started with the custom poses path.

This is stupid because a PoseInfo has a bool that says if a pose is
custom or not.

Besides that, the method of checking fails if the user ever decides to
move the game somewhere else.