Support KC and preset body swapping
In regards to preset body swapping, the method of identifying when the
body is changing needed to be changed. The new method notifies right at
the heart of maid prop processing.

The new method also removes the need for coroutines which is nice.

Some bugs were fixed as well like gravity control breaking when
switching body.