Rework gizmo
Gizmo scaling was broken and was also very finicky. Scaling just didn't
work because it relied on the gizmo's own scale to work properly. I
accidentally set the scale back to Vector3.one every time the gizmo was
dragged so scaling was broken.

Scaling also required the gizmo to be in local mode to even show up
properly. Instead the gizmo will conform to scaling modes needs rather
than keeping track of the previous state.

Checking if the gizmo was dragged was also extracted from setting the
target's transform since they're not even related to each other.