Add numerical text field gui control
The text field used with sliders is a useful GUI control that can be
used in other controls as well. The text field had some issues though.
One prominent issue is poor handling of the decimal separator.

Given the example text field value '1.05', erasing the '5' will result
in the text field formatting itself to '1'.

The NumericalTextField only formats the text when the value is set
externally so the text field does not eagerly change the text while
you're typing.