Fix sudden rotation jump when focusing on object
This issue was caused by applying the camera rotation directly. Doing
this bypasses the camera's method of applying rotations and results in a
sudden jump in rotation when focusing because the camera's internal
state was never updated.

This fix applies the rotations through the camera's methods before
overriding the z rotation.