Add focusing on maid face or body
This adds a button to the maid switcher portrait to focus on the maid's
face or body depending on the current active pane. Just like props,
double clicking while holding the select hotkey (by default A) focuses
on the face when clicking the head or the body when clicking the body or
maid cube.

Focusing on the face will move the camera directly in front of the face
so that you can see the face better while adjusting face sliders.

Focusing on the body will move the camera to the maid like with props
but try to have the maid fit within the view.