Fix PartsEdit compatibility regression
PartsEdit adds compatibility for MultipleMaids spawned props by checking
for root gameobjects with a name that ends with ".menu" and has an
animation component.

72c2e6c introduced a regression where a number is appended to the name
of the gameobject if there are duplicate props spawned.

This change applies the name of the object to the prop drag handle's
name property rather than the object's name field.