Add KeyboardShortcut and KeyboardHotkey
KeyboardShortcut is literally just BepInEx's KeyboardShortcut turned
into a class for polymorphism and a key pool.

The key pool can be set to reduce the number of keys to check.
(Although, O(n) == O(n + m)).

The main addition is the KeyboardHotkey. KeyboardHotkey functions
similarly to KeyboardShortcut but is only concerned about all the keys
in the shortcut being pressed.

For example, Z + LeftControl will only trigger for KeyboardShortcut when
Z is pressed while LeftControl is already held. So, pushing Z and then
holding LeftControl will not fire the shortcut. KeyboardHotkey doesn't
care about order.