Add DragHandle
This is an attempt to decouple 'DragPoint' from MPS and make it a
standalone UI gizmo like Custom Maid's 'GizmoRender'.

DragHandle itself functions similarly to DragPoint. The Translation
operations have been reworked to be 1:1 to the mouse cursor in contrast
to DragPoint where the drag point would lag behind the mouse cursor. The
behaviour is comparable to Unity editor's UI translation gizmo.

Builder provides a fluent interface for building a drag handle. This is
a million times better than the DragPoint.Make -> DragPoint.Initialize
-> DragPoint.Set dance with all the temporal coupling badness that comes
with it.

ClickHandler handles clicks (duh) rather than making every single drag
handle handle clicks. This allows for more complex ways of resolving
which drag handle was clicked without tanking performance per click. For
example, the click handler can sort drag handles by priority and
distance. Therefore a drag handle will be selected if it has a high
priority and it is closest to the camera.