Rework drag handle plane translation logic
The plane will now be created right at the point where the drag handle
was clicked.

The previous implementation created the plane at the same position the
drag handle's transform is located. This works fine when the drag
handles position is the same as the target's position but not when the
drag handle is offset from the target's position. When the drag handle
is offset, the translation of the target feels off.

The drag handle will also only focus on moving the target. The previous
implementation would move the target as well as the drag handle but
doing so is unnecessary since the position delegate will reposition the
drag handle anyway.