Allow drag handle to not have a target
One use case for a drag handle is to just receive mouse events so a
controller could do some transformations that do not require actually
translating the target.

This change locks the movement type of the drag handle to none when the
drag handle does not have a target. This change also cleans up null
checks for target where it would not be reachable while the drag
handle's movement type is none.