Return to AllProcPropSeqStart to detect maid load
The previous version of this used a hacky method of figuring out the
start of maid loading that I later found did not guarantee which
MaidProps are actually loading.

I found that when loading a onepiece clothing, both wear and skirt
MaidProps are detected as "loading" as well so 2 loads are detected in
total before the end of the sequence happens.

Reverting back to this method will break preset body load provided by
Perdition-117/COM3D2.EditBodyLoadFix again. I have submitted a pull
request that fixes the issue.