Save background visibility with scene
Forgot to add this when changing responsibility of background visibility
from lights to background.

Since background visibility was only affected in legacy scenes when
"colour" mode was used, lights need to load after background to override
changes to the background visibility when loading legacy scenes.

BackgroundSchema doesn't need a version bump since the whole rewrite is
a work in progress.