Add transform watcher
Before, any changes made to the transform that need to be communicated
to other objects would have to also call a method on the controller to
notify of the change. This would end up becoming a hassle to maintain
because I'd have to remember to call the "notify others" method after
having done any transform changes.

Now, a watcher will watch for any transform changes and call a callback
that the controller sets. This simplifies the controller API so that
transform changes are not done in two steps anymore.