Rework drag handle drag and double click
Drag handle used to handle mouse dragging on it's own but this should be
handle by the click handler instead. Now, the drag handle's core logic
for moving with the mouse is moved into the Drag method.

Double click used to work by just checking if the handle is clicked on
within the double click sensitivity time period regardless of the mouse
cursor's position on the second click. Now, the second click must be
within a certain distance from the first click to be registered as a
double click.