Rework gizmo to handle events like drag handle
Previously drag handle controllers would subscribe directly to gizmos
events and would do things, aside from transforming its target,
regardless of the current mode of the drag handle controller.

Now when a gizmo is interacted with, it is now handled the same way
that a drag handle is where the logic contained within DragHandleMode
derived classes are what is run. This allows for granularity for what a
gizmo should be doing based on the current mode that was not possible
with the previous handling of gizmos.