Remove "updating" flag from BasePane
There were no mechanisms to update ui control values without raising an
event and so an "updating" flag was used to indicate that the ui control
value was just being updated and that its value should not be propagated
to the underlying controller.

There are now various SetValueWithoutNotify type methods for ui controls
so the flag is not required anymore.