Migrate to Dropdown<T>
The old DropDown implementation was a list of strings that has a loose
relationship with the collection that it represents. This loose
relationship means that a dropdown item only represents an index in the
underlying collection rather than the data that that index points to.
Another issue was that the dropdown required at least 1 item and so a
dummy item is needed when the dropdown actually contains no items.

The new Dropdown<T> implementation now has a copy of the underlying
collection. This allows for accessing the data directly from the
dropdown.

Dropdown helper was changed to make it a generic helper for drawing a
scrollable window with a list of string items rather than being tied to
either DropDown or Dropdown<T>. This change allows for Combobox to use
that generic functionality for its own needs that differ from how a
Dropdown<T> control works.