Add UI scaling
Previously, UI scaling was only implemented in SceneBrowserWindow and
SceneManagementModal.

ScreenSizeChecker checks for screen size changes (duh) and notifies
subscribers.

LazyStyle is a GUIStyle wrapper that lazily initializes the style and
caches it on use. LazyStyle also updates the font size when the screen
dimensions change.

Through LazyStyle and the object wrapper versions of the various
GUILayout controls, implementing UI scaling was a breeze and only a
handful of panes with special style needs needed some changes.