Apply menu file prop texture changes
Previously the parser treated material change and tex change as the same
when they are two different things. The parser still parses it properly
because material change and texture change both have the same data types
at the same indices. In fact, this would cause spawning problems because
the texture change information is not valid material information. Now,
material changes are parsed properly and texture changes are now also
parsed in addition.

Applying texture changes fixes an issue with some props that have
different colours not being applied when spawned.