Fix scrollbar overlapping dropdown buttons
I initially wanted to add a way to check if the scrollbar is visible to
dynamically adjust the dropdown width but the method of doing so in
IMGUI is extremely unreliable because it requires knowing the margin
size of the control before the end of the scrollview. Instead I opted to
keep the dropdown width the same regardless of if there is a scrollbar
visible or not.

Pairs of sliders were adjusted as well to not overflow the width of the
window.

The main window's minimum width has also been increased slightly to
prevent any overflow past the window width.

I hate IMGUI so much. It drives me insane. I'll migrate the UI to Unity
UI some day.