Add virtual scrolling to dropdown
Virtual scrolling is an optimization that renders a portion of a list of
items that would be visible within the scroll view. The game would
suffer from huge frame drops when the amount of items in a dropdown is
very large. A prominent example of this would be the Ero 2 category of
poses with around 3000 items.