Fix achievement showing up when taking screenshot
No, taking a screenshot does not give you a custom maid achievement ;p.
The achievement panel's alpha is changed to animate fade in and out.
When a screenshot is taken, all ngui uipanels have their alpha set to 0
and then set back to 1 after. This means that even ui panels that are
invisible to start with are also visible afterwards too.