Add search bars
Things that can be searched for are:

* Base game and custom animations
* Hand presets
* Base game and custom blend sets
* Backgrounds
* All prop types

All items are searched for regardless of the currently selected
category. Animations and blend sets will search within the currently
selected type. For example if the currently selected source is custom
animations, then only custom animations will be searched.

Menu file prop filters will affect the search in a similar way.