Fix textfield forcing scroll bar
The max width text fields were causing vertical scroll bars to appear
despite there being enough space for the content. The cause seems to be
that the text within the textfield overflows the text box because the
content is larger. To fix this, just set the height of the textfield to
be larger than the containing text.