Make legacy odogu prop conversions more stable
Previously "other" props were created on demand when loading a scene
rather than getting the prop model from the repository. This was done
because an "other" prop could not be reliably determined to actually be
one. I now know that legacy odogu props were one of "other", desk and
PhotoBgObject types. This means that conversion can be reliably done and
the "other" prop repository is guaranteed (hopefully) to contain the
legacy odogu prop if it were not either a PhotoBgObject or desk prop.

BackgroundPropModels are also now retrieved from the repository. Don't
know why this wasn't done before.