Fix maid ijiri interference with face shapekeys
Maid ijiri seems to require an explicit call to TMorph.SetBlendValues
because otherwise there will be no effect when shapekeys are changed.

My reasoning for why an explicit call is required is that Maid.FaceAnime
indirectly calls TMorph.MulBlendValues which sets the blend values in
the EXACT same way as TMorph.SetBlendValues but without calling that
method directly. Essentially TMorph.SetBlendValues logic is inlined in
TMorph.MulBlendValues. Given that the logic is the exact same and
nothing happens, maid ijiri is possibly hooking onto
TMorph.SetBlendValues.

The solution is to then re-apply the blend values using
TMorph.SetBlendValues after applying a blend set.