Add SafeMpn
The MPN enum changes very often where new values are inserted in the
middle. This causes plugins compiled with older versions of MPN to
reference the wrong MPN value when the game is updated. SafeMpn is my
attempt at accessing the MPN enum in a version resilient way that is
still type safe.