Add repeat button
Repeat button is like a regular button but will constantly raise an
event as the button is held.

This implementation has a threshold before it starts repeatedly sending
signals in order to allow for just clicking the button once. An interval
can be passed to throttle how often events are sent.

Floor height and animation stepper now use repeat buttons.