Restore gravity state after character processing
When a character has any of its parts changed in edit mode, hair and
clothing gravity are reinitialized and effectively reset. This change
will restore the state of the gravity controls after processing is
finished. GravityTransformControl must be disabled because the gravity
effect is additive which means the gravity effect will stop for a moment
and then start again. I do not know if the gravity effect is
deterministic so the skirt and hair being in the same positions is not
guaranteed.