Fix free-look not applying on first load
When loading a game save file fresh, maid objects are created for the
first time and has some post-initialization steps that runs on
LateUpdate. One of the steps is to set the look target to the camera.
Post-initialization for a CharacterController happens before LateUpdate
so the free look setting is overridden.

MPS does similar post initialization tasks so the ones done by the maid
should be bypassed.

Hopefully m_MotionLoad is not used by anyone else.