Initialize services before raising called event
CalledCharacters was used for both initializing some services, like
character drag handles, and updating UI state or performing some action
after characters were called. This is problematic because the services
are not guaranteed to be initialized depending on the chain of events.

This change adds an internal event that is raised when characters are
safe to access and before the character call has completed. This allows
for essential services to be initialized and ready for access by
consumers that need to perform actions when the characters are called.