Change logic of CharacterService.Busy
Previously, Busy returned true when calling was set to false and when
none of the CharacterController's were busy. This would cause issues
similar to the CalledCharacters event where the other services were not
initialized despite the CharacterController being initialized. Now, Busy
will only return true when the characters, and any services that require
fully initialized characters, are initialized.

Interfacing with MeidoPhotoStudio should ideally be done through the
service's various events but consumers could instead be polling the Busy
property to ensure safe access.