Reset more maid things on deactivation
In the office screen, deactivation "destroys" the maid and so many
things would have been reset at that point. In edit mode the edit maid
cannot be deactivated this way and so many changes made in MPS are still
kept. This change hopefully resets the edit maid back to a neutral
state.