Add framework for extending scene data
The framework comes in 3 parts:

An extension data interface is what will be serialized to the scene and
contains an ExtensionID property that is used as the discriminator for
deserialization.

An extension data builder creates the data to be serialized and an
extension scene aspect loader will apply changes to the scene with that
data.

An extension data converter deserializes IExtensionData through a type
registry that maps extension IDs to a type that implements
IExtensionData.

All this is tied together in an extension scene aspect controller that
declares the type of the extension data and implements building and
loading extension data.