Bump plugin version to 1.1.0
Despite the version previously having a pre-release suffix according to
semver, BepInEx 5 does not use semantic versioning and so the true
version of the plugin is 1.0.0.

This is an issue because an extension system has been added to MPS and
so extensions would require a version of MPS that has this extension
system. If semver were used then dropping the pre-release suffix would
indicate that 1.0.0 > 1.0.0-beta.5.1 and the version dependency will be
met. Since this is not the case, the version has been bumped to 1.1.0
instead.