Only set Active false when mps has deactivated
When attempting to deactivate MPS, a dialogue box will ask for
confirmation to exit and Active would be set to false. This was done
in order to hide the UI in the past but today this is not needed.
Setting Active to false at this point would lock away most of the API
in a confusing manner.

Now Active will accurately represent MPS' active state.