Fix limb position discrepancy with IK
FABRIK can have limb's local position shifted while solving. The
shifting is fixed by restoring the initial positions while solving.
Non-uniform scale transform hierarchy exacerbate the issue.

Prior to this change, the solver's own backup of the local position was
used to restore the position. The problem is the backups will be
re-initialized when changing the solver's IK chain. This means that
limbs will slowly shift as the solver's chain is being changed while
posing.

The solution is to then keep and use our own backup of the local
positions.