Fix sub-header spacing preventing interaction
The text of a sub-header doesn't actually match up with the area that is
actually clickable. This is most noticeable when the header shares
horizontal space with another control and is separated from each other
with a FlexibleSpace layout control.

This could be fixed in sub-header but a toggle's clickable area should
expand past its content anyway and so this should be the way these
controls should be laid out.