Fix chest becoming smaller when clothing changes
This prominently happens when loading a scene where the chest jiggle
bone has been disabled.

The cause of this scale issue is twofold; the jiggle bone changes the
scale for a squash and stretch feature and the clothing change resets
the scale back to Vector3.one. When loading a scene, attached
accessories are detached, regardless if any are attached, which results
in the aforementioned scale reset.