Fix prop attachments issues on scene load
The issue seems to stem from the animation time being changed without
sampling the animation. The initial fix tried was to set the animation
playing state before setting the animation time which did fix the issue.
Instead of doing that I opted for fixing the problem at the source to
avoid strange behaviour when setting both those properties in the
'wrong' order.