Fix prop drag handle enabled not loading properly
Drag handle schema was intended to be used generally for drag handles
but really was only used for prop drag handles. An earlier change to
prop drag handles made the drag handle enabled property a no-op in order
to provide special functionality. This change meant that the intended
enabled property was not being saved properly.