Fix menu files not initializing with mod only enabled
Skipping game menu files when configured to do so should still wait for
the game menu database because hand items and attached items are part of
the game database. Skipping waiting when the configuration option for
ModItemsOnly is enabled would break loading all menu files so the game's
menu database must be waited for.