Rework LoadOptions to a more generic structure
Previously load options were a record struct that contained booleans and
other record structs to dictate what aspects of a scene gets loaded.
This change turns LoadOptions into a tree like structure with individual
LoadOption nodes that represent the associated aspect and potential
sub-options.

This work was done in order to allow for future work on adding load
options for extensions.