Rework extensions to require registration
Extensions were initially supposed to just use MeidoPhotoStudio's API
and if they wanted to integrate with scenes, they registered an
ExtensionSceneAspectController.

The way things work now is that extensions derive from the Extension
class and implement the GUID and Name properties and optionally an
SceneAspectController property. This Extension class allows for future
extension (heh) when required without breaking older extensions. This
also simplifies registration by only having to register an Extension
object and MPS will handle hooking up the SceneAspectController and any
other things that could be added in the future.