Add startup preset service
A startup preset is essentially a scene that will be applied when
activating/starting MPS. Previously MPS would make changes to the
current scene on startup like changing the background to the theatre,
changing the camera settings and resetting the main light. All these
changes are now stored in a default preset that the startup preset
service will load. MPS will also try to restore the previous state
before activating MPS on deactivation.

The startup preset service can be configured to not load a preset at all
and just keep the current state which is useful for those that have a
scene capture preset that is applied automatically. A custom startup
preset can also be applied instead of the default one.