Allow for unloading the edit maid
MeidoPhotoStudio is usable in the daily/office scene and edit scene.
Functionality is similar between the two but edit mode has a notable
limitation where the edit maid cannot be unloaded. This means that the
edit maid is locked at the first slot and scenes that do not have the
edit maid or have the edit maid in a different slot will not apply the
scene in the expected way. This change removes this limitation and now
matches behaviour with the daily/office scene.

On startup of MPS, the original edit maid will be called and selected as
usual but now the edit maid can be deselected and unloaded. Unloading
the edit maid will close all the edit UI panels and disable the UI from
interaction.

Other improvements to edit mode integration have been made as well
including:
- Updating the item and slider panels content to match the new edit maid
- Updating the availability of parts types based on the maid's head type