Add safe enum source generator
This generates properties for every member of a given enum that will
provide the value by parsing the enum value's name at runtime.

The source generator will add an attribute that takes an enum type
argument and must be applied to a static partial class. The source
generator will then add properties for each enum value.

The source generator will rename enum members that start with set_/get_
by changing the case of the first letter. set_/get_ are not allowed to
be used as names for properties when properties without those prefixes
already exist. For example, the enum MPN contains members called 'body'
and 'set_body' and generating both these properties unchanged would be a
compilation error. The source generator will rename properties that
start with set_/get_ even if there would not be any conflicting
properties.