Fix sepia tone/fog being applied when exiting MPS
This only really happens when having previous exited studio mode as far
as I know. What happens is that studio mode destroys the fog and sepia
tone effects on the main camera when exiting. When MPS is started,
missing effects are added but are mistakenly seen as being enabled for
the purposes of restoring state prior to MPS being activated. When MPS
is exited, the missing effects will then be re-enabled.

MPS will now check if the effects are missing and add them in a disabled
state.

The camera dependency is also pulled up into the EffectControllerBase
class.