Initialize rotation limits on access
RootMotion IKSolver.Bone sets a flag for when rotation limits components
could not be found in the IK chain. This means limits are never applied
while solving even after having added rotation limit components to the
IK chain.

This problem has been avoided in the past because the IK chain is set to
empty before the real chain. Without emptying the chain, a new chain of
the same length will reuse the existing Bone object without clearing the
rotation limit check flag.

This change is primarily for classes that would own an instance of an
IK solver and does not change the chain.