Fix scene management thumbnail hiding buttons
This was observed with scene thumbnails that had a fullscreen aspect
ratio rather than a widescreen one. The thumbnail would be as tall as
the window and the scene management buttons would be pushed out of the
window preventing any interaction including closing the modal window.

This change instead calculates a size for the thumbnail that would allow
for room for the filename, management buttons and load options toggle.